requirements.txt
# SQL Analyzer Enterprise - Production Requirements
# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# HTTP and Networking
requests==2.31.0
urllib3==2.0.7

# System Monitoring
psutil==5.9.5

# Production Server
gunicorn==21.2.0

# Environment Management
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.24.4

# File Processing
openpyxl==3.1.2
xlsxwriter==3.1.9

# JSON and Data Formats
jsonschema==4.20.0
pyyaml==6.0.1

# Date and Time
python-dateutil==2.8.2

# Logging and Monitoring
structlog==23.2.0

# Security
cryptography==41.0.8

# Performance
cachetools==5.3.2

# Development Tools (optional)
black==23.12.0
flake8==6.1.0
pytest==7.4.3
